Get rid of region interface and just use an abstract class
